Reorganizing profile, remove unused modules
