+++ profile_bluesky/startup/instrument/devices/__init__.py
@@ -1,3 +1,4 @@
 from .stages import *
 from .xspress3 import *
 from .misc_devices import *
+from .fpga_flyer import *
