+++ profile_bluesky/startup/instrument/plans/__init__.py
@@ -1,3 +1,2 @@
-from .hitp_scans import *
 from .helpers import *
-from .adapt_opt import *
+from .cont_scan import *
